feat: remove pandas dependency feat: add sentry integration
