+++ .github/workflows/release.yml
@@ -30,15 +30,6 @@ jobs:
         run: ./decrypt_credentials.sh
         env:
           CREDENTIALS_PASSWORD: ${{ secrets.CREDENTIALS_PASSWORD }}
-      # Setup Sentry release
-#      - name: Create Sentry release
-#        uses: getsentry/action-release@v1
-#        env:
-#          SENTRY_AUTH_TOKEN: ${{ secrets.SENTRY_AUTH_TOKEN }}
-#          SENTRY_ORG: ${{ secrets.SENTRY_ORG }}
-#          SENTRY_PROJECT: ${{ secrets.SENTRY_PROJECT }}
-#        with:
-#          environment: production
       # Build using SAM
       - run: sam build --use-container
       # Deploy on AWS
