+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
-name = "strapi_synchronizer"
-version = "0.1.3"
+name = "community_synchronizer"
+version = "0.2.0"
 description = ""
 authors = ["Your Name <adam@culturepulse.ai>"]
 
@@ -11,7 +11,6 @@ python-dotenv = "^0.21.0"
 pydantic = "1.10.1"
 pygsheets = "^2.0.5"
 numpy = "1.23.4"
-pandas = "^1.5.2"
 wheel = "^0.38.4"
 sentry-sdk = "^1.11.1"
